order_service/requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
certifi==2025.4.26
click==8.2.0
exceptiongroup==1.3.0
fast-depends==2.4.12
fastapi==0.115.12
faststream==0.5.41
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.4.4
pamqp==3.3.0
propcache==0.3.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
yarl==1.20.0
